requirements.txt.txt

streamlit
tensorflow
Pillow
numpy
matplotlib
seaborn
opencv-python-headless
